requirements.txt
streamlit
Pillow
openai
transformers
torch
requests